Limit purge num to 100

+++ commands/purge.js
@@ -36,6 +36,14 @@ class PurgeCommand extends Command {
             return await message.reply("Invalid number of messages to purge.")
         }
 
+        // If purge num is over 100, stop.
+        if (numToPurge > 100)
+        {
+            return await message.reply("You cannot purge any more than 100 messages at a time.")
+        }
+
+        if (numToPurge === 100) numToPurge-- // We do this to make sure numToPurge stays at 100 when it's incremented below, since apparently API only limits you to 100 messages
+
         // Add 1 to numToPurge to account for message that triggered this command
         // Automatically filter out messages that are too old to delete (Can only bulk delete messages under 14 days old)
         await message.channel.bulkDelete(numToPurge + 1, true)
